Organized GPUs by generational PCI ID ranges. (#38)
This now makes the scripts more compact as there are conditional statements
selecting the vGPU profile enabled card by the generation, and the generation
is determined by the ranges of PCI IDs. This should eliminate the need for
adding PCI IDs manually now, and open up more profile options for users to pick
from, including profiles with a VRAM multiple of 3GB.

Note that this breaks compatibility with previous versions as users may have
to go back and reselect a new profile.